@@ -10,35 +10,35 @@
     "lint:fix": "next lint --fix"
   },
   "dependencies": {
-    "@babel/core": "^7.20.2",
-    "@chakra-ui/icons": "^2.0.12",
-    "@emotion/react": "^11.10.5",
-    "@emotion/styled": "^11.10.5",
-    "chakra-react-select": "^4.4.0",
-    "core-js": "^3.26.1",
-    "firebase": "^9.14.0",
+    "@babel/core": "^7.21.3",
+    "core-js": "^3.29.1",
+    "firebase": "^9.19.0",
     "formik": "^2.2.9",
-    "framer-motion": "^6.5.1",
     "lodash": "^4.17.21",
-    "next": "12.2.3",
+    "next": "13.2.4",
+    "nookies": "^2.5.2",
     "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-firebase-hooks": "^5.1.1",
-    "swr": "^1.3.0"
+    "react-dom": "18.2.0"
   },
   "devDependencies": {
-    "@chakra-ui/react": "^2.4.1",
-    "@chakra-ui/system": "^2.3.3",
-    "@chakra-ui/theme-tools": "^2.0.13",
-    "@types/lodash": "^4.14.189",
+    "@firebase/app-types": "^0.9.0",
+    "@tailwindcss/typography": "^0.5.9",
+    "@types/lodash": "^4.14.192",
     "@types/node": "18.6.1",
     "@types/react": "18.0.15",
     "@types/react-dom": "18.0.6",
+    "autoprefixer": "^10.4.14",
+    "encoding": "^0.1.13",
     "eslint": "8.20.0",
-    "eslint-config-next": "12.2.3",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-react": "^7.31.11",
+    "eslint-config-next": "13.1.2",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-react": "^7.32.2",
     "eslint-plugin-unused-imports": "^2.0.0",
+    "firebase-admin": "^11.5.0",
+    "install": "^0.13.0",
+    "newsapi": "^2.4.1",
+    "postcss": "^8.4.21",
+    "tailwindcss": "^3.3.1",
     "typescript": "4.7.4"
   }
 }
